🧼 Separated tests related to standard library extensions
