+++ MythicWeatherTests/MythicWeatherTests.swift
@@ -133,21 +133,6 @@ class MythicWeatherTests: XCTestCase {
                                                                     ])
                                                         ])
     
-    func test_CLLocationCoordinate2DEquatableConformance() throws {
-        /// Given
-        let baseCoordinate = CLLocationCoordinate2D(latitude: 0, longitude: 0)
-        
-        /// When
-        let identicalToBaseCoordinate = CLLocationCoordinate2D(latitude: 0, longitude: 0)
-        let differentToBaseCoordinate = CLLocationCoordinate2D(latitude: 1, longitude: 0)
-        let differentToBaseCoordinateAlt = CLLocationCoordinate2D(latitude: 0, longitude: 1)
-        
-        /// Then
-        XCTAssertEqual(baseCoordinate, identicalToBaseCoordinate)
-        XCTAssertNotEqual(baseCoordinate, differentToBaseCoordinate)
-        XCTAssertNotEqual(baseCoordinate, differentToBaseCoordinateAlt)
-    }
-    
     func test_decodeWeatherFromJSON() throws {
         /// Given `expectedWeather` and `jsonResponse`
         let jsonWeatherObjetData = jsonWeatherObjet.data(using: .utf8)!
